code refactors and completed itinerary API

+++ src/services/packageList.js
@@ -1,39 +1,86 @@
 import asyncHandler from "express-async-handler";
 import PackageList from "../models/packageList.js";
+import Trip from "../models/trip.js";
 
-const create = asyncHandler(async (req, res) => {
+// old - http://localhost:5000/package-lists
+// new - http://localhost:5000/package-lists/:tripId
+export const create = asyncHandler(async (req, res) => {
   const { name } = req.body;
-  const { userId } = req.user;
-  const packageList = await PackageList.create({ name, user: userId });
+
+  const trip = await Trip.findOne({
+    _id: req.params.tripId,
+    user: req.user.userId,
+  });
+
+  if (!trip) {
+    res.status(404);
+    throw new Error("Trip not found");
+  }
+
+  const packageList = await PackageList.create({ name, trip: trip._id });
   res.status(201).json(packageList);
 });
 
-const findAll = asyncHandler(async (req, res) => {
-  const packageLists = await PackageList.find({
+export const findAll = asyncHandler(async (req, res) => {
+  const trip = await Trip.findOne({
+    _id: req.params.tripId,
     user: req.user.userId,
   });
+
+  if (!trip) {
+    res.status(404);
+    throw new Error("Trip not found");
+  }
+
+  const packageLists = await PackageList.find({
+    trip: req.params.tripId,
+  });
+
   res.status(200).json(packageLists);
 });
 
-const findById = asyncHandler(async (req, res) => {
-  const { id } = req.params;
-  const packageList = await PackageList.findOne({
-    _id: id,
+export const findById = asyncHandler(async (req, res) => {
+  const { id, tripId } = req.params;
+
+  const trip = await Trip.findOne({
+    _id: tripId,
     user: req.user.userId,
   });
-  if (!packageList) {
-    res.status(404).json({ message: "Package list not found" });
-    return;
+
+  if (!trip) {
+    res.status(404);
+    throw new Error("Trip not found");
   }
+
+  const packageList = await PackageList.findOne({
+    _id: id,
+    trip: tripId,
+  });
+
+  if (!packageList) {
+    res.status(404);
+    throw new Error("Package list not found");
+  }
+
   res.status(200).json(packageList);
 });
 
-const update = asyncHandler(async (req, res) => {
-  const { id } = req.params;
+export const update = asyncHandler(async (req, res) => {
+  const { id, tripId } = req.params;
+  const trip = await Trip.findOne({
+    _id: tripId,
+    user: req.user.userId,
+  });
+
+  if (!trip) {
+    res.status(404);
+    throw new Error("Trip not found");
+  }
+
   const packageList = await PackageList.findOneAndUpdate(
     {
       _id: id,
-      user: req.user.userId,
+      trip: tripId,
     },
     {
       ...(req.body.name && { name: req.body.name }),
@@ -41,24 +88,35 @@ const update = asyncHandler(async (req, res) => {
     },
     { new: true }
   );
+
   if (!packageList) {
-    res.status(404).json({ message: "Package list not found" });
-    return;
+    res.status(404);
+    throw new Error("Package list not found");
   }
   res.status(200).json(packageList);
 });
 
-const remove = asyncHandler(async (req, res) => {
-  const { id } = req.params;
-  const packageList = await PackageList.findOneAndDelete({
-    _id: id,
+export const remove = asyncHandler(async (req, res) => {
+  const { id, tripId } = req.params;
+  const trip = await Trip.findOne({
+    _id: tripId,
     user: req.user.userId,
   });
-  if (!packageList) {
-    res.status(404).json({ message: "Package list not found" });
-    return;
+
+  if (!trip) {
+    res.status(404);
+    throw new Error("Trip not found");
   }
+
+  const packageList = await PackageList.findOneAndDelete({
+    _id: id,
+    trip: tripId,
+  });
+
+  if (!packageList) {
+    res.status(404);
+    throw new Error("Package list not found");
+  }
+
   res.status(200).json({ message: "Package list deleted" });
 });
-
-export { create, findAll, findById, update, remove };
